requirements.txt
Flask
mysql-connector-python
matplotlib
pandas
